fixed errors in customer after merge

+++ src/main/java/com/mcnz/customers/CustomerServiceImpl.java
@@ -27,7 +27,7 @@ public class CustomerServiceImpl implements CustomerService {
 	
 	@Override
 	public void updateCustomer(Customer newCustomer) {
-		newCustomer.setId((long) 0);
+		newCustomer.setId(0);
 		customerDAO.updateCustomer(newCustomer);
 	}
